+++ app/src/main/java/nextstep/signup/MainActivity.kt
@@ -21,6 +21,15 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import nextstep.signup.components.DefaultButton
+import nextstep.signup.components.EmailTextField
+import nextstep.signup.components.HeadLine
+import nextstep.signup.components.PasswordTextField
+import nextstep.signup.components.UsernameTextField
+import nextstep.signup.state.EmailState
+import nextstep.signup.state.PasswordConfirmState
+import nextstep.signup.state.PasswordState
+import nextstep.signup.state.UsernameState
 import nextstep.signup.ui.theme.SignupTheme
 
 class MainActivity : ComponentActivity() {
