+++ app/src/main/java/nextstep/signup/components/Headlines.kt
@@ -1,4 +1,4 @@
-package nextstep.signup
+package nextstep.signup.components
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
