+++ app/src/main/java/nextstep/signup/state/InputValidationResult.kt
@@ -1,4 +1,4 @@
-package nextstep.signup
+package nextstep.signup.state
 
 enum class InputValidationResult {
     Valid,
