+++ app/src/main/java/nextstep/signup/state/UsernameState.kt
@@ -1,9 +1,9 @@
-package nextstep.signup
+package nextstep.signup.state
 
-import nextstep.signup.InputValidationResult.Empty
-import nextstep.signup.InputValidationResult.LengthError
-import nextstep.signup.InputValidationResult.Valid
-import nextstep.signup.InputValidationResult.WrongFormat
+import nextstep.signup.state.InputValidationResult.Empty
+import nextstep.signup.state.InputValidationResult.LengthError
+import nextstep.signup.state.InputValidationResult.Valid
+import nextstep.signup.state.InputValidationResult.WrongFormat
 
 data class UsernameState(
     val username: String = ""
